feat: Bigtable authorized views - move the code over to the TabularApiSurface class (#1463)
* Move the constructor over to TabularApiService

* Move sampleRowKeys over

* Move sampleRowKeys functions over and use promisify

* Adjust the proxyquire to work with TabularAPIserv

* Move all the ReadRows functionality over

* Solve the issue with the is dependency

* 🦉 Updates from OwlBot post-processor

See https://github.com/googleapis/repo-automation-bots/blob/main/packages/owl-bot/README.md

* Add header for new class

* Only include Table mocks that are necessary

* Remove TODO

* Rename TabularApiService to TabularApiSurface

* Change all imports to tabular-api-surface

* surface. not service

---------
